requirements.txt
pyelftools 

threadpoolctl 

pefile 

numpy 

joblib 

capstone 

scipy 

scikit-learn